Refactor Individual and Funder users
