+++ src/main/kotlin/com/terraformation/backend/customer/db/UserStore.kt
@@ -765,7 +765,8 @@ class UserStore(
         usersRow.cookiesConsentedTime,
         usersRow.locale,
         usersRow.timeZone,
-    )
+        parentStore,
+        permissionStore)
   }
 
   private fun insertKeycloakUser(
