+++ src/main/kotlin/com/terraformation/backend/customer/model/DeviceManagerUser.kt
@@ -57,8 +57,6 @@ data class DeviceManagerUser(
   override val defaultPermission: Boolean
     get() = false
 
-  override fun hasAnyAdminRole(): Boolean = false
-
   override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
     return mutableSetOf()
   }
