+++ src/test/kotlin/com/terraformation/backend/RunsAsDatabaseUser.kt
@@ -83,7 +83,8 @@ interface RunsAsDatabaseUser : RunsAsUser {
                     record.cookiesConsentedTime,
                     record.locale,
                     record.timeZone,
-                )
+                    ParentStore(dslContext),
+                    PermissionStore(dslContext))
           }
 
       CurrentUserHolder.setCurrentUser(user)
